[msbuild] Unify the EmbedProvisionProfile and EmbedMobileProvision tasks between iOS and macOS.
These two tasks did essentially the same thing, so we can just merge them. I
kep tthe "EmbedProvisionProfile" name, because that sounded like the most
applicable to all platforms.
