+++ msbuild/Xamarin.iOS.Tasks.Core/Xamarin.iOS.Common.targets
@@ -160,7 +160,7 @@ Copyright (C) 2013-2016 Xamarin. All rights reserved.
 
 	<PropertyGroup>
 		<_CodesignAppBundleDependsOn>
-			_EmbedMobileProvision;
+			_EmbedProvisionProfile;
 			_CodesignNativeLibraries;
 			_CollectFrameworks;
 			_CodesignFrameworks;
@@ -887,18 +887,6 @@ Copyright (C) 2013-2016 Xamarin. All rights reserved.
 		<ItemGroup>
 			<FileWrites Include="$(_TextureAtlasCache)" />
 		</ItemGroup>
-	</Target>
-
-	<Target Name="_EmbedMobileProvision" Condition="'$(_RequireCodeSigning)' == 'true'" DependsOnTargets="_GenerateBundleName;_DetectSigningIdentity"
-		Outputs="$(_AppBundlePath)embedded.mobileprovision">
-		<EmbedMobileProvision
-			SessionId="$(BuildSessionId)"
-			Condition="'$(IsMacEnabled)' == 'true'"
-			AppBundleDir="$(AppBundleDir)"
-			ProvisioningProfile="$(_ProvisioningProfile)"
-			SdkPlatform="$(_SdkPlatform)"
-			>
-		</EmbedMobileProvision>
 	</Target>
 
 	<Target Name="_DetectDebugNetworkConfiguration" Condition="'$(_BundlerDebug)' == 'true'">
