--- msbuild/Xamarin.iOS.Tasks/Tasks/EmbedMobileProvision.cs
@@ -1,6 +1,0 @@
-﻿namespace Xamarin.iOS.Tasks
-{
-	public class EmbedMobileProvision : EmbedMobileProvisionTaskBase
-	{
-	}
-}
